+++ Scripts/Handler.js
@@ -89,9 +89,6 @@ inline function onComboBox1Control(component, value)
 	}
 };
 
-
-
-
 // Animations
 const var button_animationToggle = Content.getComponent("button_animationToggle");
 button_animationToggle.setControlCallback(onbutton_animationToggleControl);
@@ -223,7 +220,6 @@ inline function onbutton_preset_rightArrowControl(component, value)
 // SHORTCUT HANDLER
 const var panel_background = Content.getComponent("panel_background");
 
-
 panel_background.setKeyPressCallback(onBackgroundKeypress);
 inline function onBackgroundKeypress(key) {
 	
@@ -280,125 +276,23 @@ inline function onbutton_x1Control(component, value)
 	showMain();
 };
 
-// REVERB
-const var Reverb = Synth.getEffect("Simple Reverb1");
-
-// BYPASS
-Content.getComponent("button_toggle_reverb").setControlCallback(onbutton_toggle_reverbControl);
-const var panel_reverb = Content.getComponent("themeablePanel_reverb");
-inline function onbutton_toggle_reverbControl(component, value)
-{
-	//Reverb.setBypassed(!value);
-	panel_reverb.set('enabled', value);
-	
-	button_freeze.setValue(!value);
-	updateFreezeParameter(false);
-};
-
-Content.getComponent("knob_reverb_space").setControlCallback(onknob_reverb_spaceControl);
-inline function onknob_reverb_spaceControl(component, value)
-{
-	//Reverb.setAttribute(Reverb.RoomSize, value);
-	updateParameterWithLabel('SPACE', value, '%');
-	showTempScreen('reverb');
-	
-	ReverbAnimationPanel.setValue(component.get("max") + 1 - (value * -1) * 7);
-	ReverbAnimationPanel.repaint();
-};
-
-Content.getComponent("knob_reverb_damping").setControlCallback(onknob_reverb_dampingControl);
-inline function onknob_reverb_dampingControl(component, value)
-{
-	//Reverb.setAttribute(Reverb.Damping, value);
-	updateParameterWithLabel('DAMPING', value, '%');
-	showTempScreen('reverb');
-	
-	AN_STATE.outterThickness = (value + 0.2) * 10 ;
-	ReverbAnimationPanel.repaint();
-};
-
-Content.getComponent("knob_reverb_stereo").setControlCallback(onknob_reverb_stereoControl);
-inline function onknob_reverb_stereoControl(component, value)
-{
-	//Reverb.setAttribute(Reverb.Width, value);
-	updateParameterWithLabel('STEREO', value, '%');
-	showTempScreen('reverb');
-	
-	AN_STATE.limit = value * 14;
-	ReverbAnimationPanel.repaint();
-};
-
-const var button_freeze = Content.getComponent("button_freeze");
-button_freeze.setControlCallback(onbutton_freezeControl);
-inline function onbutton_freezeControl(component, value)
-{
-	Reverb.setAttribute(Reverb.FreezeMode, value);
-	updateFreezeParameter(value);
-};
-
-Content.getComponent("knob_reverb_drywet").setControlCallback(onknob_reverb_drywetControl);
-inline function onknob_reverb_drywetControl(component, value)
-{
-	//Reverb.setAttribute(Reverb.WetLevel, value);
-	updateParameterWithLabel('REVERB WET', value, '%');
-	showTempScreen('reverb');
-	
-	AN_STATE.innerThickness = value * 5;
-	ReverbAnimationPanel.repaint();
-};
-
 // DEGRADE
-//const var Degrade = Synth.getEffect("Degrade1");
-// BYPASS
-Content.getComponent("button_toggle_degrade").setControlCallback(onbutton_toggle_degradeControl);
-Content.getComponent("knob_degrade_bit").setControlCallback(onknob_degrade_bitControl);
-
-inline function onknob_degrade_bitControl(component, value)
-{
-	//Degrade.setAttribute(Degrade.Quant, value);
-	updateParameterWithBit('CRUSH', value);
-	showTempScreen('degrade');
-
-	DEGRADE_STATE.innerThickness = 8 + value * -7;
-	DegradeAnimationPanel.repaint();
-};
-
-Content.getComponent("knob_degrade_rate").setControlCallback(onknob_degrade_rateControl);
-const rateMin = 0.52;
-const rateMax = 0.64;
-const rateRange = rateMax - rateMin;
-
-
-inline function onknob_degrade_rateControl(component, value)
-{
-	//Degrade.setAttribute(Degrade.Rate, value);
-
-	updateParameterWithFixedSampleRate('SAMPLE RATE', value);
-	showTempScreen('degrade');
-	
-	local rateNormalized = (value - rateMin) / rateRange;
-
-	DegradeAnimationPanel.setValue(component.get("max") + 1 - rateNormalized * 20);
-	DegradeAnimationPanel.repaint();
-};
-
 Content.getComponent("knob_degrade_mix").setControlCallback(onknob_degrade_mixControl);
 inline function onknob_degrade_mixControl(component, value)
 {
 	//Degrade.setAttribute(Degrade.PostFilt, value);
-	updateParameterWithLabel('POST FILTER', value, '%');
+	updateParameterWithLabel('DEGRADE', value, '%');
 	showTempScreen('degrade');
+	
+	
+	DegradeAnimationPanel.setValue(component.get("max") + 1 - value * 20);
 	
 	DEGRADE_STATE.corner = 80 + (value * -80);
 	DegradeAnimationPanel.repaint();
 };
 
 // FLAIR
-
-// FLAIR BYPASS
-Content.getComponent("button_toggle_flair").setControlCallback(onbutton_toggle_flairControl);
-
-Content.getComponent("knob_flair_flair").setControlCallback(onknob_flair_flairControl);
+Content.getComponent("knob_effects_flair").setControlCallback(onknob_flair_flairControl);
 inline function onknob_flair_flairControl(component, value)
 {
 	//Flair.setAttribute(Flair.WetAmount, value);
@@ -410,160 +304,6 @@ inline function onknob_flair_flairControl(component, value)
 	
 	Flair1AnimationPanel.set('x', initialPosX + move);
 	Flair2AnimationPanel.set('x', initialPosX - move);
-};
-
-// FILTER
-const var Filter = Synth.getEffect("Parametriq EQ1");
-const var panel_filterButtons = Content.getComponent("panel_filterButtons");
-
-// BYPASS
-Content.getComponent("button_toggle_filter").setControlCallback(onbutton_toggle_filterControl);
-const var panel_filter = Content.getComponent("themeablePanel_filter");
-
-inline function onbutton_toggle_filterControl(component, value)
-{
-	Filter.setBypassed(!value);
-	panel_filter.set('enabled', value);
-	
-	if (STATE.presetBrowserOpen) return;
-	
-	if (value) {
-		STATE.filterOpen = true;
-		displayShowMain('filter');
-		updateFreezeParameter(false);
-	} else {
-		STATE.filterOpen = false;
-		displayShowMain('default');
-	}
-	
-	button_showFilter.setValue(value);
-	
-};
-
-
-const var knob_filter_freq = Content.getComponent("knob_filter_freq");
-knob_filter_freq.setControlCallback(onknob_filter_freqControl);
-inline function onknob_filter_freqControl(component, value)
-{
-	Filter.setAttribute(STATE.currentBandIndex + 1 , value);
-	updateParameterWithLabel('FREQUENCY', value, 'Hz');
-	
-	if (STATE.filterOnDrag == 1) {
-		STATE.filterOpen = true;
-		displayShowMain('filter');
-		updateFreezeParameter(false);
-	}
-};
-
-const var knob_filter_q = Content.getComponent("knob_filter_q");
-knob_filter_q.setControlCallback(onknob_filter_qControl);
-inline function onknob_filter_qControl(component, value)
-{
-	
-	Filter.setAttribute(STATE.currentBandIndex + 2, value);
-	updateParameterWithLabel('Q', value, '');
-	if (STATE.filterOnDrag == 1) {
-		STATE.filterOpen = true;
-		displayShowMain('filter');
-		updateFreezeParameter(false);
-	}
-}
-
-const var knob_filter_gain = Content.getComponent("knob_filter_gain");
-knob_filter_gain.setControlCallback(onknob_filter_gainControl);
-inline function onknob_filter_gainControl(component, value)
-{
-	Filter.setAttribute(STATE.currentBandIndex, value);
-	updateParameterWithDb('GAIN', value);
-	if (STATE.filterOnDrag == 1) {
-		STATE.filterOpen = true;
-		displayShowMain('filter');
-		updateFreezeParameter(false);
-	}
-};
-
-const var button_showFilter = Content.getComponent("button_showFilter");
-button_showFilter.setControlCallback(onbutton_showFilterControl);
-inline function onbutton_showFilterControl(component, value)
-{
-	if (value) {
-		STATE.filterOpen = true;
-		displayShowMain('filter');
-		updateFreezeParameter(false);
-	} else {
-		STATE.filterOpen = false;
-		displayShowMain('default');
-	}
-};
-
-Content.getComponent("button_disableBand").setControlCallback(onbutton_disableBandControl);
-inline function onbutton_disableBandControl(component, value)
-{	
-	Filter.setAttribute(STATE.currentBandIndex + 3, value);
-	label_bandDisplay.set('enabled', value);
-};
-
-// Display Filter Knobs
-const filterButtons = [
-	Content.getComponent("button_lowPass"),
-	Content.getComponent("button_HighPass"),
-	Content.getComponent("button_lowShelf"),
-	Content.getComponent("button_highShelf"),
-	Content.getComponent("button_bandPass")
-];
-
-filterButtons[0].setControlCallback(onbutton_lowPassControl);
-inline function onbutton_lowPassControl(component, value)
-{
-	
-	if (value) {
-			local idx = 0;
-			Filter.setAttribute(STATE.currentBandIndex + 4, idx);
-			filterTypeRadio(idx);
-	}
-};
-
-filterButtons[1].setControlCallback(onbutton_HighPassControl);
-inline function onbutton_HighPassControl(component, value)
-{
-	
-	if (value) {
-		local idx = 1;
-		Filter.setAttribute(STATE.currentBandIndex + 4, idx);
-		filterTypeRadio(idx);
-	}
-};
-
-filterButtons[2].setControlCallback(onbutton_lowShelfControl);
-inline function onbutton_lowShelfControl(component, value)
-{
-	
-	if (value) {
-		local idx = 2;
-		Filter.setAttribute(STATE.currentBandIndex + 4, idx);
-		filterTypeRadio(idx);
-	
-	}
-};
-
-filterButtons[3].setControlCallback(onbutton_highShelfControl);
-inline function onbutton_highShelfControl(component, value)
-{
-	if (value) {
-		local idx = 3;
-		Filter.setAttribute(STATE.currentBandIndex + 4, idx);
-		filterTypeRadio(idx);
-	}
-};
-
-filterButtons[4].setControlCallback(onbutton_bandPassControl);
-inline function onbutton_bandPassControl(component, value)
-{
-	local idx = 4;
-	if (value) {
-		Filter.setAttribute(STATE.currentBandIndex + 4, idx);
-		filterTypeRadio(idx);
-	}
 };
 
 // MASTER
