--- Scripts/ScriptProcessors/Reach/Flair.js
@@ -1,13 +1,0 @@
- function prepareToPlay(sampleRate, blockSize)
-{
-	
-}
- function processBlock(channels)
-{
-	
-}
- function onControl(number, value)
-{
-	
-}
-
