+++ XmlPresetBackups/Reach.xml
@@ -9,36 +9,34 @@
         <Processor Type="ScriptProcessor" ID="Interface" Bypassed="0">
           <ChildProcessors/>
           <Content>
-            <Control type="ScriptSlider" id="header_vumeter_ch1" value="-0.11"/>
-            <Control type="ScriptSlider" id="header_vumeter_ch2" value="-0.11"/>
+            <Control type="ScriptSlider" id="header_vumeter_ch1" value="-0.1099999994039536"/>
+            <Control type="ScriptSlider" id="header_vumeter_ch2" value="-0.1099999994039536"/>
             <Control type="ScriptButton" id="Button1" value="0.0"/>
-            <Control type="ScriptSlider" id="knob_reverb_space" value="0.0"/>
+            <Control type="ScriptSlider" id="knob_reverb_space" value="0.5699999928474426"/>
+            <Control type="ScriptSlider" id="knob_reverb_diffusion" value="0.0"/>
             <Control type="ScriptSlider" id="knob_reverb_damping" value="0.0"/>
-            <Control type="ScriptSlider" id="knob_reverb_stereo" value="0.0"/>
-            <Control type="ScriptSlider" id="knob_reverb_drywet" value="0.0"/>
-            <Control type="ScriptButton" id="button_freeze" value="1"/>
-            <Control type="ScriptSlider" id="knob_degrade_bit" value="0.0"/>
-            <Control type="ScriptSlider" id="knob_degrade_rate" value="0.0"/>
+            <Control type="ScriptSlider" id="knob_reverb_time" value="0.0"/>
+            <Control type="ScriptButton" id="button_freeze" value="0.0"/>
+            <Control type="ScriptSlider" id="knob_effects_flair" value="0.0"/>
             <Control type="ScriptSlider" id="knob_degrade_mix" value="0.0"/>
             <Control type="ScriptSlider" id="knob_filter_gain" value="0.0"/>
             <Control type="ScriptSlider" id="knob_filter_freq" value="0.0"/>
             <Control type="ScriptSlider" id="knob_filter_q" value="0.0"/>
             <Control type="ScriptSlider" id="knob_io_in" value="0.0"/>
-            <Control type="ScriptSlider" id="knob_io_out" value="1.0"/>
+            <Control type="ScriptSlider" id="knob_io_out" value="0.2199999988079071"/>
             <Control type="ScriptButton" id="button_not_activated" value="0.0"/>
             <Control type="ScriptButton" id="button_website" value="0.0"/>
             <Control type="ScriptButton" id="button_github" value="0.0"/>
-            <Control type="ScriptButton" id="button_toggle_reverb" value="0.0"/>
-            <Control type="ScriptButton" id="button_toggle_degrade" value="0.0"/>
-            <Control type="ScriptButton" id="button_toggle_flair" value="0.0"/>
-            <Control type="ScriptButton" id="button_toggle_filter" value="0.0"/>
+            <Control type="ScriptButton" id="button_toggle_reverb" value="1.0"/>
+            <Control type="ScriptButton" id="button_toggle_filter" value="1.0"/>
             <Control type="ScriptButton" id="button_disableBand" value="0.0"/>
             <Control type="ScriptButton" id="button_lowPass" value="0.0"/>
             <Control type="ScriptButton" id="button_HighPass" value="0.0"/>
             <Control type="ScriptButton" id="button_lowShelf" value="0.0"/>
             <Control type="ScriptButton" id="button_highShelf" value="0.0"/>
             <Control type="ScriptButton" id="button_bandPass" value="0.0"/>
-            <Control type="ScriptSlider" id="knob_flair_flair" value="0.0"/>
+            <Control type="ScriptSlider" id="knob_modulation_depth" value="0.0"/>
+            <Control type="ScriptSlider" id="knob_modulation_frequency" value="0.0"/>
           </Content>
           <UIData Source="ReachUIData"/>
         </Processor>
@@ -57,8 +55,8 @@
           <RoutingMatrix NumSourceChannels="4" Channel0="0" Send0="2" Channel1="1" Send1="3"
                          Channel2="-1" Send2="-1" Channel3="-1" Send3="-1"/>
         </Processor>
-        <Processor Type="SimpleGain" ID="DryGain" Bypassed="0" Gain="-100.0" Delay="0.0"
-                   Width="100.0" Balance="0.0" InvertPolarity="0.0">
+        <Processor Type="SimpleGain" ID="DryGain" Bypassed="0" Gain="-2.158108234405518"
+                   Delay="0.0" Width="100.0" Balance="0.0" InvertPolarity="0.0">
           <ChildProcessors>
             <Processor Type="ModulatorChain" ID="Gain Modulation" Bypassed="0" Intensity="1.0">
               <ChildProcessors/>
@@ -76,8 +74,8 @@
           <RoutingMatrix NumSourceChannels="4" Channel0="0" Send0="-1" Channel1="1" Send1="-1"
                          Channel2="-1" Send2="-1" Channel3="-1" Send3="-1"/>
         </Processor>
-        <Processor Type="SimpleGain" ID="WetGain" Bypassed="0" Gain="0.0" Delay="0.0"
-                   Width="100.0" Balance="0.0" InvertPolarity="0.0">
+        <Processor Type="SimpleGain" ID="WetGain" Bypassed="0" Gain="-13.15154647827148"
+                   Delay="0.0" Width="100.0" Balance="0.0" InvertPolarity="0.0">
           <ChildProcessors>
             <Processor Type="ModulatorChain" ID="Gain Modulation" Bypassed="0" Intensity="1.0">
               <ChildProcessors/>
@@ -95,36 +93,26 @@
           <RoutingMatrix NumSourceChannels="4" Channel0="-1" Send0="-1" Channel1="-1"
                          Send1="-1" Channel2="2" Send2="-1" Channel3="3" Send3="-1"/>
         </Processor>
-        <Processor Type="ScriptFX" ID="FaustReverb" Bypassed="0">
+        <Processor Type="Hardcoded Master FX" ID="FaustReverb" Bypassed="0" Network="FaustReverb"
+                   Damping="0.3599999845027924" Diffusion="0.7070000171661377" HF Gain="1.0"
+                   LF Gain="1.0" Mod Depth="0.1000000014901161" Mod Frequency="0.1000000014901161"
+                   ReverbTime="12.0" Size="5.0">
           <ChildProcessors/>
           <RoutingMatrix NumSourceChannels="4" Channel0="-1" Send0="-1" Channel1="-1"
                          Send1="-1" Channel2="2" Send2="-1" Channel3="3" Send3="-1"/>
-          <Content/>
-          <Networks>
-            <Network ID="Reverb" AllowCompilation="1" Version="0.0.0"/>
-          </Networks>
-          <UIData>
-            <ContentProperties DeviceType="Desktop"/>
-          </UIData>
         </Processor>
-        <Processor Type="SimpleReverb" ID="Simple Reverb1" Bypassed="0" RoomSize="0.7699999809265137"
+        <Processor Type="SimpleReverb" ID="Freeverb" Bypassed="0" RoomSize="0.7699999809265137"
                    Damping="0.0" WetLevel="1.0" DryLevel="0.0" Width="0.6399999856948853"
-                   FreezeMode="1.0">
+                   FreezeMode="0.0">
           <ChildProcessors/>
           <RoutingMatrix NumSourceChannels="4" Channel0="-1" Send0="-1" Channel1="-1"
                          Send1="-1" Channel2="2" Send2="-1" Channel3="3" Send3="-1"/>
         </Processor>
-        <Processor Type="ScriptFX" ID="Flair" Bypassed="0">
+        <Processor Type="Hardcoded Master FX" ID="Flair" Bypassed="0" Network="Flair"
+                   Mix="1.0" Depth="1.0">
           <ChildProcessors/>
           <RoutingMatrix NumSourceChannels="4" Channel0="-1" Send0="-1" Channel1="-1"
                          Send1="-1" Channel2="2" Send2="-1" Channel3="3" Send3="-1"/>
-          <Content/>
-          <Networks>
-            <Network ID="Flair" Version="0.0.0" AllowCompilation="1"/>
-          </Networks>
-          <UIData>
-            <ContentProperties DeviceType="Desktop"/>
-          </UIData>
         </Processor>
         <Processor Type="ScriptFX" ID="Degrade" Bypassed="0">
           <ChildProcessors/>
@@ -138,7 +126,7 @@
             <ContentProperties DeviceType="Desktop"/>
           </UIData>
         </Processor>
-        <Processor Type="CurveEq" ID="Parametriq EQ1" Bypassed="1" NumFilters="1"
+        <Processor Type="CurveEq" ID="Parametriq EQ1" Bypassed="0" NumFilters="1"
                    Band0="0.0" Band1="20.0" Band2="0.300000011920929" Band3="0.0"
                    Band4="1.0" FFTEnabled="1">
           <ChildProcessors/>
