+++ settings.json
@@ -2,5 +2,5 @@
   "zoom": 0.6000000238418579,
   "animationEnabled": 1,
   "filterOnDrag": 1,
-  "theme": "Light"
+  "theme": "Dark"
 }
